# doc/swagger/v1/swagger.yml
openapi: 3.0.3
info:
  title: Article Management System
  description: >-
    Collection of tests for the article management system API.


    [https://github.com/nathanbizkit/article-management](https://github.com/nathanbizkit/article-management)
  version: 1.0.0
  contact: {}
servers:
  - url: http://localhost:8000/api/v1
  - url: https://localhost:8443/api/v1
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: session
    refreshAuth:
      type: apiKey
      in: cookie
      name: refreshToken
security:
  - sessionAuth: []
paths:
...
  /articles:
    get:
      tags:
        - Articles
      summary: All Global Articles
      description: Retrieves all global articles.
      operationId: allGlobalArticles
      parameters:
        - name: tag
          in: query
          schema:
            type: string
        - name: username
          description: Author's username
          in: query
          schema:
            type: string
        - name: favorited
          description: Username of whom you want to see their favorited articles
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: number
        - name: offset
          in: query
          schema:
            type: number
      responses:
        "200":
          description: List of article objects
          content:
            application/json:
              schema:
                type: object
                properties:
                  articles:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: number
                        title:
                          type: string
                        description:
                          type: string
                        body:
                          type: string
                        tags:
                          type: array
                          items:
                            type: string
                        favorited:
                          type: boolean
                        favorites_count:
                          type: number
                        author:
                          type: object
                          properties:
                            username:
                              type: string
                            name:
                              type: string
                            bio:
                              type: string
                            image:
                              type: string
                              format: uri
                            following:
                              type: boolean
                        created_at:
                          type: string
                          format: date-time
                        updated_at:
                          type: string
                          format: date-time
                  articles_count:
                    type: number
    post:
      tags:
        - Articles
      summary: Create Article
      description: Creates a new article.
      operationId: createArticle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                body:
                  type: string
                description:
                  type: string
                tags:
                  type: array
                  items:
                    type: string
      responses:
        "200":
          description: An article object
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: number
                  title:
                    type: string
                  description:
                    type: string
                  body:
                    type: string
                  tags:
                    type: array
                    items:
                      type: string
                  favorited:
                    type: boolean
                  favorites_count:
                    type: number
                  author:
                    type: object
                    properties:
                      username:
                        type: string
                      name:
                        type: string
                      bio:
                        type: string
                      image:
                        type: string
                        format: uri
                      following:
                        type: boolean
                  created_at:
                    type: string
                    format: date-time
                  updated_at:
                    type: string
                    format: date-time
...
  /articles/{slug}:
    get:
      tags:
        - Articles
      summary: Article by Slug
      description: Retrieves an article by slug.
      operationId: articleBySlug
      responses:
        "200":
          description: An article object
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: number
                  title:
                    type: string
                  description:
                    type: string
                  body:
                    type: string
                  tags:
                    type: array
                    items:
                      type: string
                  favorited:
                    type: boolean
                  favorites_count:
                    type: number
                  author:
                    type: object
                    properties:
                      username:
                        type: string
                      name:
                        type: string
                      bio:
                        type: string
                      image:
                        type: string
                        format: uri
                      following:
                        type: boolean
                  created_at:
                    type: string
                    format: date-time
                  updated_at:
                    type: string
                    format: date-time
    put:
      tags:
        - Articles
      summary: Update Article by Slug
      description: Updates an article.
      operationId: updateArticle
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
                body:
                  type: string
      responses:
        "200":
          description: An article object
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: number
                  title:
                    type: string
                  description:
                    type: string
                  body:
                    type: string
                  tags:
                    type: array
                    items:
                      type: string
                  favorited:
                    type: boolean
                  favorites_count:
                    type: number
                  author:
                    type: object
                    properties:
                      username:
                        type: string
                      name:
                        type: string
                      bio:
                        type: string
                      image:
                        type: string
                        format: uri
                      following:
                        type: boolean
                  created_at:
                    type: string
                    format: date-time
                  updated_at:
                    type: string
                    format: date-time
    delete:
      tags:
        - Articles
      summary: Delete Article by Slug
      description: Deletes an article by slug.
      operationId: deleteArticle
      responses:
        "204":
          description: Successfully deleted the article.
    parameters:
      - name: slug
        description: Article's id
        in: path
        required: true
        schema:
          type: number
  /articles/{slug}/favorite:
    post:
      tags:
        - Articles
      summary: Favorite Article by Slug
      description: Favorites an article by slug.
      operationId: favoriteArticle
      responses:
        "200":
          description: An article object
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: number
                  title:
                    type: string
                  description:
                    type: string
                  body:
                    type: string
                  tags:
                    type: array
                    items:
                      type: string
                  favorited:
                    type: boolean
                  favorites_count:
                    type: number
                  author:
                    type: object
                    properties:
                      username:
                        type: string
                      name:
                        type: string
                      bio:
                        type: string
                      image:
                        type: string
                        format: uri
                      following:
                        type: boolean
                  created_at:
                    type: string
                    format: date-time
                  updated_at:
                    type: string
                    format: date-time
    delete:
      tags:
        - Articles
      summary: Unfavorite Article by Slug
      description: Unfavorites an article by slug.
      operationId: unfavoriteArticle
      responses:
        "200":
          description: An article object
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: number
                  title:
                    type: string
                  description:
                    type: string
                  body:
                    type: string
                  tags:
                    type: array
                    items:
                      type: string
                  favorited:
                    type: boolean
                  favorites_count:
                    type: number
                  author:
                    type: object
                    properties:
                      username:
                        type: string
                      name:
                        type: string
                      bio:
                        type: string
                      image:
                        type: string
                        format: uri
                      following:
                        type: boolean
                  created_at:
                    type: string
                    format: date-time
                  updated_at:
                    type: string
                    format: date-time
    parameters:
      - name: slug
        description: Article's id
        in: path
        required: true
        schema:
          type: number
...